+++ R/create_env.R
@@ -13,7 +13,7 @@
 #'
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #'   create_env(system.file("extdata", "testing_environment.yml", package = "DALEXtra"))
 #' }
 #' @rdname create_env
@@ -29,7 +29,7 @@ create_env <- function(yml, condaenv = NULL) {
         condaenv
       ))
     } else {
-      stop("Conda not found")
+       stop("Conda not found")
     }
   }