+++ R/explain_xgboost.R
@@ -1,13 +1,13 @@
-#' Create explainer from your h2o model
+#' Create explainer from your xgboost model
 #'
 #' DALEX is designed to work with various black-box models like tree ensembles, linear models, neural networks etc.
 #' Unfortunately R packages that create such models are very inconsistent. Different tools use different interfaces to train, validate and use models.
-#' One of those tools, we would like to make more accessible is H2O.
+#' One of those tools, we would like to make more accessible is xgboost.
 #'
 #'
 #' @param model object - a model to be explained
 #' @param data data.frame or matrix - data that was used for fitting. If not provided then will be extracted from the model. Data should be passed without target column (this shall be provided as the \code{y} argument). NOTE: If target variable is present in the \code{data}, some of the functionalities my not work properly.
-#' @param y numeric vector with outputs / scores. If provided then it shall have the same size as \code{data}
+#' @param y numeric vector with outputs / scores. If provided then it shall have the same size as \code{data}. For classif task has to be numerci in range [0, nclasses)
 #' @param weights numeric vector with sampling weights. By default it's \code{NULL}. If provided then it shall have the same length as \code{data}
 #' @param predict_function function that takes two arguments: model and new data and returns numeric vector with predictions
 #' @param residual_function function that takes three arguments: model, data and response vector y. It should return a numeric vector with model residuals for given data. If not provided, response residuals (\eqn{y-\hat{y}}) are calculated.
@@ -18,6 +18,8 @@
 #' @param colorize if TRUE (default) then \code{WARNINGS}, \code{ERRORS} and \code{NOTES} are colorized. Will work only in the R console.
 #' @param model_info a named list (\code{package}, \code{version}, \code{type}) containg information about model. If \code{NULL}, \code{DALEX} will seek for information on it's own.
 #' @param type type of a model, either \code{classification} or \code{regression}. If not specified then \code{type} will be extracted from \code{model_info}.
+#' @param encode_function fuction(data, ...) that if executed with \code{data} parameters returns encoded dataframe that was used to fit model. Xgboost does not handle factors on it's own so such function is needed to aquire better explanations.
+#' @param true_labels vecotr of \code{y} before encoding.
 #'
 #' @return explainer object (\code{\link[DALEX]{explain}}) ready to work with DALEX
 #'
@@ -26,63 +28,30 @@
 #'
 #'
 #' @examples
-#' \donttest{
+#' library("xgboost")
+#' library("DALEXtra")
+#' library("mlr")
+#' # 8th column is target that has to be omitted in X data
+#' data <- as.matrix(createDummyFeatures(titanic_imputed[,-8]))
+#' model <- xgboost(data, titanic_imputed$survived, nrounds = 10,
+#'                  params = list(objective = "binary:logistic"),
+#'                 prediction = TRUE)
+#' # explainer with encode functiom
+#' explainer_1 <- explain_xgboost(model, data = titanic_imputed[,-8],
+#'                                titanic_imputed$survived,
+#'                                encode_function = function(data) {
+#'  as.matrix(createDummyFeatures(data))
+#' })
+#' plot(predict_parts(explainer_1, titanic_imputed[1,-8]))
 #'
+#' # explainer without encode function
+#' explainer_2 <- explain_xgboost(model, data = data, titanic_imputed$survived)
+#' plot(predict_parts(explainer_2, data[1,,drop = FALSE]))
 #'
-#' # # load packages and data
-#' library(h2o)
-#' library(DALEXtra)
-#'
-#' # data <- DALEX::titanic_imputed
-#'
-#' # init h2o
-#'  h2o.init()
-#'
-#' # stop h2o progress printing
-#'  h2o.no_progress()
-#'
-#' # split the data
-#' # h2o_split <- h2o.splitFrame(as.h2o(data))
-#' # train <- h2o_split[[1]]
-#' # test <- as.data.frame(h2o_split[[2]])
-#' # h2o automl takes target as factor
-#' # train$survived <- as.factor(train$survived)
-#'
-#' # fit a model
-#' # automl <- h2o.automl(y = "survived",
-#' #                   training_frame = train,
-#' #                    max_runtime_secs = 30)
-#'
-#'
-#' # create an explainer for the model
-#' # explainer <- explain_h2o(automl,
-#' #                        data = test,
-#' #                         y = test$survived,
-#' #                          label = "h2o")
-#'
-#'
-#' titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
-#' titanic_train <- read.csv(system.file("extdata", "titanic_train.csv", package = "DALEXtra"))
-#' titanic_h2o <- h2o::as.h2o(titanic_train)
-#' titanic_h2o["survived"] <- h2o::as.factor(titanic_h2o["survived"])
-#' titanic_test_h2o <- h2o::as.h2o(titanic_test)
-#' model <- h2o::h2o.gbm(
-#' training_frame = titanic_h2o,
-#' y = "survived",
-#' distribution = "bernoulli",
-#' ntrees = 500,
-#' max_depth = 4,
-#' min_rows =  12,
-#' learn_rate = 0.001
-#' )
-#' explain_h2o(model, titanic_test[,1:17], titanic_test[,18])
-#'
-#' h2o.shutdown(prompt = FALSE)
-#' }
-#' @rdname explain_h2o
+#' @rdname explain_xgboost
 #' @export
 
-explain_h2o <-
+explain_xgboost <-
   function(model,
            data = NULL,
            y = NULL,
@@ -95,19 +64,14 @@ explain_h2o <-
            precalculate = TRUE,
            colorize = TRUE,
            model_info = NULL,
-           type = NULL) {
+           type = NULL,
+           encode_function = NULL,
+           true_labels = NULL) {
 
-    if (class(y) == "H2OFrame") {
-      y <- as.vector(y)
-    }
-
-    if (class(model) == "H2OAutoML") {
-      message("`model` argument is a H2OAutoML class object. model@leader will be extracted.\n")
-      model <- model@leader
-    }
-
-    if (class(data) == "H2OFrame") {
-      data <- as.data.frame(data)
+    attr(model, "encoder") <- encode_function
+    if (!is.null(true_labels)) {
+      attr(model, "true_labels") <- true_labels
+      y <- true_labels
     }
 
 
@@ -126,4 +90,5 @@ explain_h2o <-
       model_info = model_info,
       type = type
     )
+
   }